requirements.txt
wheel
pre-commit
numpy
scikit-learn
pandas
pyyaml
yellowbrick
tsfresh
dvc
# dvc-gdrive

# For Ubuntu 20.04
# Also needed OS related dependencies:
# sudo apt-get install gcc libpq-dev -y
# sudo apt-get install python-dev  python-pip -y
# sudo apt-get install python3-dev python3-pip python3-venv python3-wheel -y
# pip3 install wheel
# Ubuntu 22.04 seems to work fine without this. 